python/first-infra/step1/pyproject.toml
[project]
name = "step1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.171.1",
]
